Restore guint64 to rejection variable

+++ src/stacking/median_and_mean.c
@@ -451,7 +451,7 @@ static void norm_to_0_1_range(fits *fit) {
  * does a different operation to keep the final pixel values.
  *********************************************************************************/
 
-static int percentile_clipping(WORD pixel, float sig[], float median, gint rej[]) {
+static int percentile_clipping(WORD pixel, float sig[], float median, guint64 rej[]) {
 	float plow = sig[0];
 	float phigh = sig[1];
 
@@ -469,7 +469,7 @@ static int percentile_clipping(WORD pixel, float sig[], float median, gint rej[]
 /* Rejection of pixels, following sigma_(high/low) * sigma.
  * The function returns 0 if no rejections are required, 1 if it's a high
  * rejection and -1 for a low-rejection */
-static int sigma_clipping(WORD pixel, float sig[], float sigma, float median, gint rej[]) {
+static int sigma_clipping(WORD pixel, float sig[], float sigma, float median, guint64 rej[]) {
 	float sigmalow = sig[0];
 	float sigmahigh = sig[1];
 
@@ -491,7 +491,7 @@ static void Winsorize(WORD *pixel, WORD m0, WORD m1, int N) {
 	}
 }
 
-static int line_clipping(WORD pixel, float sig[], float sigma, int i, float a, float b, gint rej[]) {
+static int line_clipping(WORD pixel, float sig[], float sigma, int i, float a, float b, guint64 rej[]) {
 	float sigmalow = sig[0];
 	float sigmahigh = sig[1];
 
@@ -505,7 +505,7 @@ static int line_clipping(WORD pixel, float sig[], float sigma, int i, float a, f
 	return 0;
 }
 
-static int apply_rejection_ushort(struct _data_block *data, int nb_frames, struct stacking_args *args, gint crej[2]) {
+static int apply_rejection_ushort(struct _data_block *data, int nb_frames, struct stacking_args *args, guint64 crej[2]) {
 	int N = nb_frames;	// N is the number of pixels kept from the current stack
 	float median = 0.f;
 	int pixel, output, changed, n, r = 0;
@@ -671,7 +671,7 @@ static int apply_rejection_ushort(struct _data_block *data, int nb_frames, struc
 }
 
 static double mean_and_reject(struct stacking_args *args, struct _data_block *data,
-		int stack_size, data_type itype, gint crej[2]) {
+		int stack_size, data_type itype, guint64 crej[2]) {
 	double mean;
 	if (itype == DATA_USHORT) {
 		int kept_pixels = apply_rejection_ushort(data, stack_size, args, crej);
@@ -992,7 +992,7 @@ static int stack_mean_or_median(struct stacking_args *args, gboolean is_mean) {
 			size_t pdata_idx = (naxes[1] - (my_block->start_row + y) - 1) * naxes[0];
 			/* index of the line in the read data, data->pix[frame] */
 			size_t line_idx = y * naxes[0];
-			gint crej[2] = {0, 0};
+			guint64 crej[2] = {0, 0};
 			if (retval) break;
 
 			// update progress bar
